apply usages
